casi todo acabado printf

+++ print_f/pruebas/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: irifarac <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/02 13:30:47 by irifarac          #+#    #+#             */
-/*   Updated: 2022/03/03 14:12:55 by irifarac         ###   ########.fr       */
+/*   Updated: 2022/03/03 18:07:04 by irifarac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,11 +23,11 @@ int		ft_printf(const char *str, ...);
 int		ft_specifier(va_list lst, int str);
 int		ft_putchar(char c);
 int		ft_putstr(char *str);
-int		ft_putunsigned(unsigned int n);
+int		ft_putunsigned(long long int n);
 int		ft_puthex(unsigned int n);
 //Funciones de la libreria libft
 size_t	ft_strlen(char *str);
-char	*ft_itoa(size_t n, int base);
+char	*ft_itoa(long long n, int base);
 int		ft_len(size_t n, int base);
 
 #endif
